Add scaffolding for documentation
Documentation is an important part of this project, and I
would like to explore the ideas that I have developed while
working on the Live After 60 project. Although many documentation
generators exist, such as Doxygen, I have opted to use LaTeX owing
to a rich feature set provided by it in addition to the flexibility
and rigour that it entails.

The scaffolding for documentation has been prepared with a skeletal
Developer Manual now in place with the following stub sections:
  * Title page
  * Table of contents
  * List of figures
  * List of tables
  * Overview
  * Application Programming Interface
  * Implementation
  * Test Cases
  * Appendix
  * Index

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>