Define compiler extensions
Experience has shown that a few of the extensions provided by GCC
are quite useful. Wrapping them around macros provides a clean way
to degrade them on compilers that do not support the GCC extensions.

The compiler extensions that are known to be most likely used have
been defined as follows:
  * ag_likely()
  * ag_unlikely()
  * ag_pure()
  * ag_hot
  * ag_cold
  * ag_threadlocal

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>