Document compiler extensions
The Developer Manual for the Argent Library has been updated
with the documentation for the compiler extensions module. The
interface and implementation documentation for the following
macros have been included:
  * ag_cold
  * ag_hot
  * ag_likely()
  * ag_pure
  * ag_threadlocal
  * ag_unlikely()

Although initially I had considered segregating the interface
and implementation documentation into separate parts, I have
chosen to maintain the entire documentation related to a
particular macro in one place. The rationale for this is that
it makes the document much more concise and does not require
for the developer referring to it to have to keep making
context switches between different sections.

I spent much time experimenting with a layout that I am happy
with, and the documentation for the compiler extensions now
serves as a template for the documentation of the remaining
modules. I have also added appropriate sections on the license
terms.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>