Develop exception handling mechanism
The exception handling mechanism has been developed with the
following interface:
  * ag_erno enumeration
  * ag_exception struct
  * ag_exception_handler callback
  * ag_exception_message() function
  * ag_exception_handler_get() function
  * ag_exception_handler_set() function
  * ag_assert() macro
  * ag_require() macro

Initially, I had planned simply to implement the ag_assert() and
ag_require() macros without any form of exception handling, but a
bit of thought showed that it was doable without much effort.

The exception handling mechansim works by allowing the client code
to define an exception handler (through the ag_exception_handler
callback) which is then subsequently used by the Argent Library. In
case no exception handler is defined by the client code, then a
default handler is used instead.

There is an important distinction between ag_assert() and
ag_require(), corresponding to the difference between an assertion
and an exception. An assertion is used to ensure that a condition
which should never occur (resulting from a programming error) is
caught, whereas an exception is raised to indicate a condition that
may go wrong at runtime, such as heap failure. Accordingly,
ag_assert() checks may be disabled at runtime once we are sure that
there is no programming error, but runtime checks through
ag_require() are never disabled.

Additionally, I have updated the Developer Manual references for the
compiler extensions module.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>