Add memory pool interface
The memory pool interface has been developed with the following
functions:
  * ag_mempool_new()
  * ag_mempool_resize()
  * ag_mempool_free()

The current interface is not a true memory pool implementation;
instead it is simply a wrapper around the standard heap memory
functions. However, I do hope in time to develop this interface
into a true memory pool system. So even if it's a misnomer, it
serves as a reminder.

Additionally, the build rules have been updated so that the
shared Argent Library is now built. An unused parameter
warning was also noted in the implementation of the
eh_default() helper function in src/exception.c, and this
has now been fixed.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>