Introduce object v-table
The interface for the object v-table has been created with the
following functions:
  * ag_object_vtable_init()
  * ag_object_vtable_exit()
  * ag_object_vtable_exists()
  * ag_object_vtable get()
  * ag_object_vtable_set()

The object v-table interface has been implemented as a hash table
containing the methods for each registered object type. In order
to support the object v-table, the following changes have been made
to the object interface:
  * renamed ag_object_vtable struct to ag_object_method
  * added ag_object_sz()
  * added ag_object_type()
  * converted ag_object_hash() to a method

Some cosmetic changes have also been made, such as the removal of
the ag_object_payload type, the trimming down of comments and the
renaming of the disposal routines.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>