Aligned malloc not required currently
Since we're using GNU Libc for the current version, we don't
need to worry about the alignment of malloc() as it is guaranteed
 to be 8 bytes for 32-bit systems and 16 bytes for 64-bit systems.
In a future version, we could take a more portable approach using
aligned_alloc() and posix_memalign() and account for other libc-s
such as uCLibc and Musl.

However, for improved readability, it does make sense to have an
ag_memblock_t type, with the _t suffix indicating that it is a type
that should be manipulated through its API. The ag_object type has
also been renamed to ag_object_t.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>