Implement logging mechanism
The logging interface has been implemented with the following functions:
  * ag_log_emerg()
  * ag_log_alert()
  * ag_log_crit()
  * ag_log_err()
  * ag_log_warning()
  * ag_log_notice()
  * ag_log_info()
  * ag_log_debug()

The logging interface is compatible with syslog, and the functions have
been named to reflect the standard syslog priority naming convention.
Since systemd is becoming increasingly prevalent, the test cases have
been written assuming that the systemd-journald.service is running.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>