Document memory block interface
The memory block interface has been documented, both in the source code
and in the LaTeX files. Experience has shown that it makes sense for the
documentation to remain close to the source code, especially since this
makes it easier to make changes. Although I will preserve the LaTeX
documentation, the preference would be to extract the relevant
documentation from the source code into the LaTeX files.

This commit has been rebased from the issue/9/memblock-doc branch.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>