Create Testing Module
The Testing Module is responsible for providing a mechanism for running
unit tests in a standardised manner, both for the Argent Library itself
and for client code. The Testing Module provides the following features:
  - hierarchical organisation of tests
  - skipping of debug tests
  - reporting of tests passed, failed and skipped
  - logging of test results

The Testing Module is organised around a hierarchy of entities, the most
fundamental being unit tests. Each unit test is wrapped within a test
case, and related test cases are collected together in a test suite. The
individual test suites are in turn collected within a test harness. Each
of these entities is reified as a type, and supported by a public
interface.

The implementation of the interfaces of this Module is divided logically
across the argent/src/test-*.c files.

Currently, the Testing Module is rough around the edges, but will be
progressively improved (along with the documentation) as other issues
are closed. An important feature that is required, but has not yet been
implemented, is that of catching the SIGABRT and SIGSEGV signals.

This pull request closes #5.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>