Improve exception handling interface
The improved exception handling interface is now in place, with the
introduction of an exception registry that can be called by client code
to set exception messages and handlers associated with error codes.
Error codes have also been streamlined, now that we are adopting the
convention that error codes related to the Argent Library are all
negative whereas those of client code are positive.

While working on improving the exception handling interface, it became
apparent that a manager interface was required for correctly
initialising and terminating the Argent Library; this is now in place.
Additionally, the improved memory block interface now uses the
AG_REQUIRE_OPT() macro in its implementation. Other minor improvements
and bug fixes have also been incorporated for the testing module, and
the existing logging module has been reused.

This commit has been rebased from the issue/40/better-exh branch.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>