Rewrite string interface
The string interface has been rewritten using the reference counting
features provided by the new memory block interface. It is hoped that
this makes the string implementation a bit cleaner and easier to
maintain since the plumbing work associated with reference counting is
taken care of by the memory block module.

The string interface has also been expanded, and the following functions
are now available:
  - ag_str_new()
  - ag_str_new_fmt()
  - ag_str_new_empty()
  - ag_str_copy()
  - ag_str_release()
  - ag_str_cmp()
  - ag_str_lt()
  - ag_str_eq()
  - ag_str_gt()
  - ag_str_len()
  - ag_str_sz()
  - ag_str_refc()
  - ag_str_has()
  - ag_str_empty()
  - ag_str_lower()
  - ag_str_upper()
  - ag_str_proper()
  - ag_str_split()
  - ag_str_split_right()

Of these functions, only ag_str_lower(), ag_str_upper() and
ag_str_proper() are *not* Unicode-safe as of now; full Unicode support
has been left for a future date. There is also a caveat regarding
whether ag_str_len() correctly counts accounts for combined glyphs as in
the case of the Devanagari script.

Along with the development of the new string module, the new memory
block interface has been  provided with the functions ag_mblock_retain()
and ag_mblock_release() to explicitly increment and decrement the
reference count. ag_mblock_release() disposes the memory block
automatically if the reference count falls to 0. It would seem that the
ag_mblock_dispose() function is no longer required, but it has been kept
for the time being; if found completely vestigial it will be removed at
a later date. Additionally, the AG_AUTO() macro has been introduced to
help simplify the declaration of pointers that are decorated with a
cleanup attribute.

The testing module now uses the revised string module, and provides
better logging of test statistics.

This commit has been rebased from the issue/44/mblock-str branch.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>