Add JSON method to object interface
Every object now supports a method to return the JSON representation of
itself through ag_object_json(). Additionally, a major improvement has
been made to the object interface. The AG_OBJECT_DECLARE() and
AG_OBJECT_DEFINE() macros have been redefined to further reduce the
amount of plumbing code. Furthermore, the object interface now only
requires those object methods which need to be overridden to be defined.
The macros which help to do this are:
  - AG_OBJECT_DEFINE_CLONE()
  - AG_OBJECT_DEFINE_RELEASE()
  - AG_OBJECT_DEFINE_CMP()
  - AG_OBJECT_DEFINE_VALID()
  - AG_OBJECT_DEFINE_SZ()
  - AG_OBJECT_DEFINE_LEN()
  - AG_OBJECT_DEFINE_HASH()
  - AG_OBJECT_DEFINE_STR()
  - AG_OBJECT_DEFINE_JSON()

The layout of the code has also been improved, with the source files now
organised into logical sub-directories. An advantage of this layout has
been that the include/ sub-directory is no longer required.

This commit has been rebased from the issue/76/json-method branch.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>